+++ tasks_handlers_categories/content/content_tk_photographer.py
@@ -32,27 +32,38 @@ async def content_tk_photographer_date_handler(query: CallbackQuery, user_data, 
 
 async def input_content_tk_photographer_date_handler(message: types.Message, user_data, **kwargs):
     global task_info
+    is_last_in_album = False
+
     if message.text:
         task_info["goods_info"] = message.text
     else:
         task_info["goods_info"] = '-'
-    #Обработка медиа в сообщении
+    
+    # Обработка медиа в сообщении
     task_info = await process_media(message, task_info)
 
-    if "last_bot_message_id" in user_data[message.from_user.id]:
-        await bot.delete_message(chat_id=message.chat.id, message_id=user_data[message.from_user.id]["last_bot_message_id"])
-        del user_data[message.from_user.id]["last_bot_message_id"]
+    # Check if the message is the last in its album
+    if not message.media_group_id or message.media_group_id != user_data.get("last_media_group_id", None):
+        is_last_in_album = True
+        user_data["last_media_group_id"] = message.media_group_id
 
-    keyboard_markup = await content_tk_photographer_goods_info_keyboard(user_data["prev_action"])
-    sent_message = await bot.send_message(
-        chat_id=message.chat.id,
-        text="Данные записаны.\nДобавьте другие файлы и/или информацию при необходимости",
-        reply_markup=keyboard_markup
-    )
-    user_data[message.from_user.id] = {
-        "current_message": "content_tk_photographer_date",
-        "last_bot_message_id": sent_message.message_id
-    }
+
+    if is_last_in_album:
+
+        if "last_bot_message_id" in user_data[message.from_user.id]:
+            await bot.delete_message(chat_id=message.chat.id, message_id=user_data[message.from_user.id]["last_bot_message_id"])
+            del user_data[message.from_user.id]["last_bot_message_id"]
+    
+        keyboard_markup = await content_tk_photographer_goods_info_keyboard(user_data["prev_action"])
+        sent_message = await bot.send_message(
+            chat_id=message.chat.id,
+            text="Данные записаны.\nДобавьте другие файлы и/или информацию при необходимости",
+            reply_markup=keyboard_markup
+        )
+        user_data[message.from_user.id] = {
+            "current_message": "content_tk_photographer_date",
+            "last_bot_message_id": sent_message.message_id
+        }
 
 #Если дат не ввели
 #Ожидание описания задачи
@@ -61,7 +72,6 @@ async def input_content_tk_photographer_description_handler_without_date(query: 
         task_info['photo_paths'] = []
     if 'document_paths' not in task_info:
         task_info['document_paths'] = []
-
     user_data[query.from_user.id] = {
         "current_message": "content_tk_photographer_description",
         "last_bot_message_id": query.message.message_id
@@ -83,12 +93,11 @@ async def input_content_tk_photographer_description_handler(message: types.Messa
 
     confirmation_message = (
         "Пожалуйста, удостоверьтесь в правильности собранных данных:\n"
-        f"⚪️ Категория задачи: {task_info['task_category']}\n"
-        f"⚪️ Подкатегория задачи: {task_info['task_subcategory']}\n"
-        f"⚪️ Данные о товарах: {task_info['goods_info']}\n"
-        f"⚪️ Описание задачи: {task_info['task_description']}"
+        f"\n⚪️ Категория задачи: {task_info['task_category']}\n"
+        f"\n⚪️ Подкатегория задачи: {task_info['task_subcategory']}\n"
+        f"\n⚪️ Данные о товарах: {task_info['goods_info']}\n"
+        f"\n⚪️ Описание задачи: {task_info['task_description']}"
     )
-
     keyboard_markup = await task_confirm_keyboard()
     await bot.send_message(
         chat_id=message.chat.id,
@@ -102,12 +111,11 @@ async def content_tk_photographer_confirmation_handler_without_description(query
     task_info["task_description"] = "-"
     confirmation_message = (
         "Пожалуйста, удостоверьтесь в правильности собранных данных:\n"
-        f"⚪️ Категория задачи: {task_info['task_category']}\n"
-        f"⚪️ Подкатегория задачи: {task_info['task_subcategory']}\n"
-        f"⚪️ Данные о товарах: {task_info['goods_info']}\n"
-        f"⚪️ Описание задачи: {task_info['task_description']}"
+        f"\n⚪️ Категория задачи: {task_info['task_category']}\n"
+        f"\n⚪️ Подкатегория задачи: {task_info['task_subcategory']}\n"
+        f"\n⚪️ Данные о товарах: {task_info['goods_info']}\n"
+        f"\n⚪️ Описание задачи: {task_info['task_description']}"
     )
-
     keyboard_markup = await task_confirm_keyboard()
     await query.message.edit_text(
         text=confirmation_message,
